Created BackupContainerFileSystem.actor.h
